infil_green_ampt.ui: Added the dthetan (normal) and dthetad (dry) labels and comboboxes
infil_editor_widget.py: Added the new comboboxes to the GreenAmptDialog init method
Added the DTHETAn (normal) and DTHETAd (dry) to the save_green_ampt_shapefile_fields method
Added the DTHETAn (normal) and DTHETAd (dry) to the restore_green_ampt_shapefile_fields method

infiltration_tools.py: Added the references to the soil_dthetad_fld and soil_dthetan_fld to the InfiltrationCalculator init method
 Added the references to the dthetan_fld="DTHETAn" and dthetad_fld="DTHETAd" to the setup_green_ampt method
 Added the if statement to verify if the saturation is normal, dry, or wet

 grid_tools.py: DeprecationWarning. A variable was changed to converted to integer

 flo2d.py: Converted two strings to raw string in order to avoid escape characters

+++ flo2d/flo2d_tools/grid_tools.py
@@ -1364,7 +1364,7 @@ def gridRegionGenerator(gutils, grid, gridSpan=100, regionPadding=50, showProgre
                 if showProgress == True:
                     if progDialog.wasCanceled() == True:
                         break
-                    progDialog.setValue(regionCounter / regionCount * 100.0)
+                    progDialog.setValue(int(regionCounter / regionCount * 100.0))
                     QApplication.processEvents()
                 print("Processing region: %s of %s" % (regionCounter, regionCount))
                 yield request
